⚡ Replace shader references with Unlit shader
I found that in the WebGL build of the game there were shaders being included and erroring that I wasn't using in the project. I investigated and found that Unity had hidden references to the Shaders and that I can replace those with another shader. Very sneaky but it would reduce the size of the build and prevent the errors from occurring.

These sorts of changes could obviously cause issues if these assets were used for any other way.

At this point I've probably done more than anyone else has ever done to make a Unity project as small and as simple as possible.